+++ src/core/components/viewers/map/src/MapSidebar/src/FileTab/src/ModelsSection.tsx
@@ -13,7 +13,6 @@ import { BimContext, BuildingsContext } from '../../../../../../../../store'
 import ConfirmDialog from '../../../../../../../ConfirmDialog'
 import { CollapsibleSection } from '../../../../../../../ui/CollapsibleSection'
 import { useFileDeleteHandler, FileItemComponent, useFileActions, useFileUploadWithProgress, UploadProgressBar, useUploadTasks } from '../../../../../../../ui/FilesManager'
-import { LoadingSpinner } from '../../../../../../../ui/LoadingSpinner'
 import { toggleBimToMap as dispatchToggleBimToMap } from '../../../../../utils/toggleBimToMap'
 
 import type { DbFile as DbFile } from '../../../../../../../../types/dbTypes'